Fix "no blank line" warning.
Fix warnings from "dzil build" from plugin PkgVersion about
"no blank line for $VERSION after package". This is to warn the user that
Dist::Zilla had to insert a new line into the given file to add the
version number. This causes line numbers to be off by one in the
generated distribution compared to line numbers in the source file.
By inserting a blank line after the package statement in the source
file this problem can be fixed (and the the warning from "dzil build"
goes away).